refactor: move `QuoteRouter` into production.
